refactor: ♻️ sharedディレクトリにパブリックAPIを追加しました
