--- packages/app/src/entities/index.ts
@@ -1,3 +1,0 @@
-export * from "./board";
-export * from "./cell";
-export * from "./player";
