+++ packages/app/src/scenes/game/ui/BoradCell.tsx
@@ -4,7 +4,7 @@ import { useGame } from "@/game/hooks/useGame";
 import { useUpdateGame } from "@/game/hooks/useUpdateGame";
 import { canClickTile, getTileState } from "@/game/logics/tileLogic";
 import type { Tile } from "@/game/types/tile";
-import { cn, preloadImages } from "@/shared/utils/helpers";
+import { cn, preloadImages } from "@/shared/utils";
 import type { FC } from "react";
 import { useTranslation } from "react-i18next";
