+++ packages/app/src/shared/libs/index.ts
@@ -1,0 +1,1 @@
+export * from "./firebase";
